Feat: metods put and delete
Inserindo os métodos de atualizar e excluir usuário nas rotas

+++ src/app/api/users/route.js
@@ -1,8 +1,10 @@
 import Users from "../../../models/users";
 import { NextResponse } from "next/server";
+import connectMongoDB from "../../../../config/database";
 
 export async function GET() {
   try {
+    connectMongoDB();
     const users = await Users.find();
 
     return NextResponse.json({ users }, { status: 200 });
@@ -16,11 +18,42 @@ export async function GET() {
 }
 
 export async function DELETE(request) {
+  connectMongoDB();
   const id = request.nextUrl.searchParams.get("id");
 
-  await Users.findByIdAndDelete(id);
-  return NextResponse.json({ message: "Usuário deletado" }, { status: 200 });
+  try {
+    const userToDelete = await Users.findById(id);
+
+    if (!userToDelete) {
+      return NextResponse.json(
+        { message: "Usuário não encontrado" },
+        { status: 404 }
+      );
+    }
+
+    const deletedUser = await Users.findByIdAndDelete(id);
+
+    return NextResponse.json(
+      { message: `Usuário ${deletedUser.nome} deletado` },
+      { status: 200 }
+    );
+  } catch (error) {
+    console.error("Erro ao excluir usuário:", error);
+    return NextResponse.json(
+      { message: "Erro ao excluir usuário" },
+      { status: 500 }
+    );
+  }
 }
+
+/* export async function DELETE(req, { params }) {
+  try {
+    const { id } = params;
+    await Users.findByIdAndDelete(id);
+    return NextResponse.json({ message: "Usuário deletado" }, { status: 200 });
+  } catch (error) {}
+  return NextResponse.json({ message: "Error", error }, { status: 500 });
+} */
 
 /* export async function GET_ID(req) {
   try {
@@ -39,6 +72,7 @@ export async function DELETE(request) {
 } */
 
 export async function POST(req) {
+  connectMongoDB();
   try {
     /* const body = await req.json();
     const usersData = body.formData;
